Implement support for REPLCONF GETACK command on replica server
This commit adds support for the REPLCONF GETACK command on the replica server. When the master sends this command, the replica server responds with REPLCONF ACK
0, indicating that it has processed 0 bytes of commands. This mechanism allows the master to keep track of the replica's progress and ensure synchronization.
Proper offset tracking will be implemented in the next stage. After the master-replica handshake, the replica should only respond to REPLCONF GETACK commands and
not send responses for other propagated commands.