Implement support for EXEC command without MULTI command
This commit adds support for the EXEC command in Redis when it is executed without having called the MULTI command. It returns an error message of "ERR EXEC
without MULTI" as a Simple error. The commit also includes updated instructions, examples, and tests for this functionality. This implementation is part of the
ongoing work to support Redis commands in the course-definition.yml file.